message_board_docker/requirements.txt
asgiref==3.4.1
astroid==2.11.2
async-generator==1.10
attrs==21.4.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.10.0
bitstring==3.1.9
blinker==1.4
boto3==1.22.1
botocore==1.25.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.6
chromedriver-autoinstaller==0.3.1
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.2
decorator==5.0.9
dill==0.3.4
Django==3.2.7
dnspython==2.2.0
docopt==0.6.2
ecdsa==0.17.0
esptool==3.1
Flask==2.0.2
Flask-Mail==0.9.1
Flask-S3==0.3.3
Flask-WTF==1.0.1
gTTS==2.2.4
h11==0.13.0
idna==3.3
invoke==1.6.0
ipython==7.26.0
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
Js2Py==0.71
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.7.0
mysql-connector-python==8.0.28
numpy==1.21.4
outcome==1.1.0
packaging==21.3
pandas==1.3.4
parso==0.8.2
pickleshare==0.7.5
pipwin==0.5.2
platformdirs==2.5.1
prompt-toolkit==3.0.19
protobuf==3.19.4
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.4.0
pygame==2.1.0
Pygments==2.9.0
pyjsparser==2.7.1
pylint==2.13.4
pymongo==4.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
PyPrind==2.11.3
pyserial==3.5
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyxel==1.4.3
reedsolo==1.5.4
regex==2021.10.23
requests==2.27.0
rich==10.10.0
s3transfer==0.5.2
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
traitlets==5.0.5
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.14.0
wsproto==1.1.0
WTForms==3.0.1
yapf==0.31.0
